requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.0.0
# Дашборд (Streamlit)
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
